+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 0.49.0 - TBD
 
 #### Enhancements
+- Added new venues, datasets, and publishers for ICE Futures US and for ICE Futures
+  Europe (Financial products)
 - Added a `keep_zip` parameter to `Historical.batch.download()`. When `True`, and downloading all files, the jobs contents will be saved as a ZIP file
 - Calling `Live.terminate()` will now attempt to write EOF before aborting the connection to help close the remote end